requirements.txt
pytest==6.1.1
selenium==3.141.0
python-dotenv==0.10.5
allure-pytest==2.9.45